requirements.txt
numpy==1.26.4
opencv_python_headless==4.8.1.78
Pillow==10.0.0
pynput==1.7.6
PySide2==5.15.2.1
